feat: add new methods and attributes to the 'Obra' classes
Add methods to count and get works by client id and work status. Add new attributes to the entity “Obra". Add methods to the service to update the work, validate the previous status of the work, finalize the work, assign a client to a work and validate if a client is enabled.